Add back Socket::pair
Moves it the sys module as its only available on Unix.
